HBX-1961: Move class 'org.hibernate.tool.internal.reveng.RevEngUtils' to 'org.hibernate.tool.internal.reveng.util.RevengUtils'
